+++ time_test.go
@@ -4,7 +4,217 @@ import (
 	"encoding/json"
 	"testing"
 	"time"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb/dynamodbattribute"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb/expression"
+	"github.com/edwardsmatt/dynamocity/internal/testutils"
 )
+
+var db *dynamodb.Client
+var tableName *string
+var itemsSortedOrder []testuils.TestDynamoItem
+var err error
+
+func init() {
+	/* load test data */
+	db, tableName, itemsSortedOrder, err = testutils.SetupTestFixtures()
+}
+
+func Test_DynamocityTime(t *testing.T) {
+	if err != nil {
+		t.Error(err)
+		t.FailNow()
+	}
+
+	cases := []SortKeyTestCase{
+		{
+			Name:       "Given a RFC3339 Timestamp, when using a sortkey with dynamocity time, then apply sort key greaterThan based on nanotime precision",
+			Timestamp:  "2019-12-09T06:50:02.53323Z",
+			SortKey:    "dynamocityTime",
+			KeyName:    "dynamocity-time-index",
+			KeyBuilder: DynamocityTimeKeyBuilder,
+			Verify: func(allItems []map[string]dynamodb.AttributeValue, tc SortKeyTestCase, t *testing.T) {
+				actualItems := make([]TestDynamoItem, len(allItems))
+				if err = dynamodbattribute.UnmarshalListOfMaps(allItems, &actualItems); err != nil {
+					t.Error(err)
+					t.FailNow()
+				}
+
+				lastFiveItems := itemsSortedOrder[5:]
+
+				if len(actualItems) != len(lastFiveItems) {
+					t.Errorf("Unexpected number of query items returned. Expected '%d', Got '%d'", len(lastFiveItems), len(actualItems))
+				}
+
+				for i, actual := range actualItems {
+					expectedItem := lastFiveItems[i]
+					if actual.SortKey != expectedItem.SortKey {
+						t.Errorf("Unexpected item ID at index %d. Expected '%s', Got '%s'", i, expectedItem.SortKey, actual.SortKey)
+					}
+					if !actual.DynamocityTime.Time().Equal(expectedItem.DynamocityTime.Time()) {
+						t.Errorf("Unexpected item time at index %d. Expected '%s', Got '%s'", i, expectedItem.DynamocityTime, actual.DynamocityTime)
+					}
+				}
+			},
+		},
+		{
+			Name:       "Given a RFC3339Nano Timestamp, when using a default RFC3339Nano timestamp, then apply sort key greaterThan based on non-precise nano precision",
+			Timestamp:  "2019-12-09T06:50:02.53323Z",
+			SortKey:    "nanoTime",
+			KeyName:    "nano-time-index",
+			KeyBuilder: NanoTimeKeyBuilder,
+			Verify: func(allItems []map[string]dynamodb.AttributeValue, tc SortKeyTestCase, t *testing.T) {
+				actualItems := make([]TestDynamoItem, len(allItems))
+				if err = dynamodbattribute.UnmarshalListOfMaps(allItems, &actualItems); err != nil {
+					t.Error(err)
+					t.FailNow()
+				}
+
+				expectedItemTimestampsInStringOrder := []string{
+					"2019-12-09T06:50:02.53323Z",
+					"2019-12-09T06:50:02.5332Z",
+					"2019-12-09T06:50:02.533Z",
+					"2019-12-09T06:50:02.53Z",
+					"2019-12-09T06:50:02.5Z",
+					"2019-12-09T06:50:02Z",
+				}
+
+				if len(actualItems) != 6 {
+					t.Errorf("Unexpected number of saved items. Expected '%d', Got '%d'", 6, len(actualItems))
+				}
+				isFirstElement := true
+				for i, actual := range actualItems {
+					if actual.StringTime != expectedItemTimestampsInStringOrder[i] {
+						t.Errorf("Unexpected String sorting order. Expected '%s', Got '%s'", expectedItemTimestampsInStringOrder[i], actual.StringTime)
+					}
+					if isFirstElement {
+						isFirstElement = false
+						continue
+					}
+					previousElement := actualItems[i-1]
+					if actual.NanoTime.After(previousElement.NanoTime) {
+						t.Errorf("Item NanoTime is unexpectedly after previous item NanoTime. This '%s', Previous '%s'", actual.NanoTime, previousElement.NanoTime)
+					}
+				}
+			},
+		},
+		{
+			Name:       "Given a RFC3339Nano Timestamp, then verify the string attribute value has truncated nanos",
+			Timestamp:  "2018-12-31T00:00:00Z",
+			SortKey:    "nanoTime",
+			KeyName:    "nano-time-index",
+			KeyBuilder: NanoTimeKeyBuilder,
+			Verify: func(allItems []map[string]dynamodb.AttributeValue, tc SortKeyTestCase, t *testing.T) {
+
+				expectedItemTimestampsInStringOrder := []string{
+					"2019-12-09T06:50:02.533237329Z",
+					"2019-12-09T06:50:02.53323732Z",
+					"2019-12-09T06:50:02.5332373Z",
+					"2019-12-09T06:50:02.533237Z",
+					"2019-12-09T06:50:02.53323Z",
+					"2019-12-09T06:50:02.5332Z",
+					"2019-12-09T06:50:02.533Z",
+					"2019-12-09T06:50:02.53Z",
+					"2019-12-09T06:50:02.5Z",
+					"2019-12-09T06:50:02Z",
+				}
+
+				if len(allItems) != 10 {
+					t.Errorf("Unexpected number of items. Expected '%d', Got '%d'", 10, len(allItems))
+				}
+
+				for i, actual := range allItems {
+					avString := actual[tc.SortKey].S
+					expectedItem := expectedItemTimestampsInStringOrder[i]
+					if *avString != expectedItem {
+						t.Errorf("Unexpected string attribute value %d. Expected '%s', Got '%s'", i, *avString, expectedItem)
+					}
+				}
+			},
+		},
+		{
+			Name:       "Given a dynamocity.Time Timestamp, then verify the string attribute value has retained nanos",
+			Timestamp:  "2018-12-31T00:00:00Z",
+			SortKey:    "dynamocityTime",
+			KeyName:    "dynamocity-time-index",
+			KeyBuilder: DynamocityTimeKeyBuilder,
+			Verify: func(allItems []map[string]dynamodb.AttributeValue, tc SortKeyTestCase, t *testing.T) {
+
+				expectedItemTimestampsInStringOrder := []string{
+					"2019-12-09T06:50:02.000000000Z",
+					"2019-12-09T06:50:02.500000000Z",
+					"2019-12-09T06:50:02.530000000Z",
+					"2019-12-09T06:50:02.533000000Z",
+					"2019-12-09T06:50:02.533200000Z",
+					"2019-12-09T06:50:02.533230000Z",
+					"2019-12-09T06:50:02.533237000Z",
+					"2019-12-09T06:50:02.533237300Z",
+					"2019-12-09T06:50:02.533237320Z",
+					"2019-12-09T06:50:02.533237329Z",
+				}
+
+				if len(allItems) != 10 {
+					t.Errorf("Unexpected number of items. Expected '%d', Got '%d'", 10, len(allItems))
+				}
+
+				for i, actual := range allItems {
+					avString := actual[tc.SortKey].S
+					expectedItem := expectedItemTimestampsInStringOrder[i]
+					if *avString != expectedItem {
+						t.Errorf("Unexpected string attribute value %d. Expected '%s', Got '%s'", i, *avString, expectedItem)
+					}
+				}
+			},
+		},
+	}
+
+	for _, tc := range cases {
+
+		lsiKeyCondition := expression.KeyAnd(
+			expression.Key("pk").Equal(expression.Value("TEST")),
+			expression.Key(tc.SortKey).GreaterThanEqual(expression.Value(tc.KeyBuilder(tc, t))),
+		)
+
+		expr, err := expression.NewBuilder().
+			WithKeyCondition(lsiKeyCondition).
+			Build()
+
+		if err != nil {
+			t.Error(err)
+			t.FailNow()
+		}
+
+		allItems := []map[string]dynamodb.AttributeValue{}
+		input := &dynamodb.QueryInput{
+			TableName:                 tableName,
+			IndexName:                 aws.String(tc.KeyName),
+			KeyConditionExpression:    expr.KeyCondition(),
+			ExpressionAttributeNames:  expr.Names(),
+			ExpressionAttributeValues: expr.Values(),
+		}
+
+		query := db.QueryRequest(input)
+		paginator := dynamodb.NewQueryPaginator(query)
+		for paginator.Next(query.Context()) {
+			page := paginator.CurrentPage()
+			allItems = append(allItems, page.Items...)
+		}
+
+		if err := paginator.Err(); err != nil {
+			t.Error(err)
+			t.FailNow()
+		}
+
+		queryResultItems := make([]TestDynamoItem, len(allItems))
+		if err = dynamodbattribute.UnmarshalListOfMaps(allItems, &queryResultItems); err != nil {
+			t.Error(err)
+			t.FailNow()
+		}
+		tc.Verify(allItems, tc, t)
+	}
+}
 
 func Test_StringSortedness(t *testing.T) {
 	sortedStrings := []string{
